Rename Markdown editor files - Fix imports
